requirements.txt
asgiref==3.2.7
Django==3.0.7
django-comments-xtd==2.6.1
django-contrib-comments==1.9.2
djangorestframework==3.11.0
docutils==0.16
flake8==3.8.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
